labs/lab-9/requirements.txt
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.34
tomli==2.0.2
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
